--- 230502/230508비밀글설정_윤서준/README.md
@@ -1,22 +1,0 @@
-# Comment-writer
-멋쟁이사자처럼 5기 쌈빡한사자들(3조)의 댓글작성기 프로젝트 레파지토리입니다
-
-## 목적
-* 자바스크립트로 배운 이론들을 댓글작성기 구현을 통해 실제로 적용해본다.
-
-* GitHub Organization을 이용해서 협업 과정을 경험해본다.
-
-## 진행방법
-- 기존의 종태님이 작성하신 댓글작성기 코드를 바탕으로 각 주차마다 설정한 추가기능들을 본인의 방식대로 구현해보고, 
-   매주 월요일 회고 스터디 시간에 짧게 해당내용을 공유한다.
-
-- 설명이 끝나면 질문이나 피드백, 혹은 코드 리뷰 시간을 갖는다.
-
-## 주차별 기능구현 내용
-
-- 2023/05.02~05.08: https://www.notion.so/c6d73e3ea15b40d3872a814f35b19111?pvs=4 
-
-
-
-
-
