--- 230502/230509비밀글설정_남종태/comment.html
@@ -1,45 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko-KR">
-<head>
-  <meta charset="UTF-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title></title>
-</head>
-<body>
-  <!-- createElement요소로 생성할 댓글 요소들 입니다. -->
-  <!-- Tip :
-    저도 얼마전에 사용해 봤는데 이 코드 그대로 복사해서 chat-gpt한테 createElement 요소로 바꿔달라하면 알아서 바꿔줍니다. 정말 편리해요 -->
-    <!-- 그냥 추가할 요소들을 태그로 직접 짜시고 chat-gpt한테 createElment 변환 시키는게 정말 좋은거 같습니다. -->
-    <li class="comment-item" id="${item.id}">
-      <div class="profile-box">
-        <div class="user-profile">
-          <img class="profile-img" src="./img/profile.png">
-          <span class="auth">${item.auth}</span>
-        </div>
-        <div class="comment-btns">
-          <!--숨김 댓글 표시 버튼-->
-        
-          <button class="show-btn"></button>
-          <button class="edit-btn"></button>
-          <button class="del-btn"></button>
-        </div>
-      </div>
-      <div class="comment-info">
-        <p class="comment-content">${item.content}</p>
-        <span class="createdAt">${getCreatedAt(item.createdAt)}</span>
-      </div>
-      <!-- 댓글 수정 폼 처음에는 보이지 않습니다. 수정시에만 보이게 처리합니다 -->
-      <form class="edit-form">
-        <textarea class="input-editComment">${item.content}</textarea>
-        <div class="textarea-footer">
-          <span class="text-counter">${item.content.length}/100</span>
-          <div class="editComment-btns">
-            <button class="cancel-btn" type="button">취소하기</button>
-            <button class="editComment-btn" type="button">수정하기</button>
-          </div>
-        </div>
-      </form>
-    </li>
-</body>
-</html>
