--- 230502/230512customPrompt 및 모듈화_남종태/.vscode/settings.json
@@ -1,3 +1,0 @@
-{
-  "liveServer.settings.port": 5501
-}
