--- 230502/230512customPrompt 및 모듈화_남종태/index.html
@@ -1,83 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko-KR">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>댓글 작성기</title>
-    <link rel="stylesheet" href="./src/style/reset.css" />
-    <link rel="stylesheet" href="./src/style/modal.css">
-    <link rel="stylesheet" href="./src/style/style.css" />
-  </head>
-  <body>
-    <header>
-      <h1 class="title">댓글 작성기</h1>
-    </header>
-    <main>
-        <h2 class="section-title">
-          <img class="comment-icon" src="./src/img/comment-icon.png" alt="댓글" />
-          댓글
-        </h2>
-        <form class="comment-form">
-          <label class="a11y-hidden" for="input-auth">작성자</label>
-          <input
-            id="input-auth"
-            type="text"
-            placeholder="작성자"
-            maxlength="8"
-          />
-          <label class="a11y-hidden" for="input-pw">비밀번호</label>
-          <input
-            id="input-pw"
-            type="password"
-            placeholder="비밀번호"
-            maxlength="16"
-          />
-          <div class="textarea-wrapper">
-            <label class="a11y-hidden" for="input-comment">내용</label>
-            <textarea
-              id="input-comment"
-              placeholder="개인정보를 공용 및 요청하거나, 명예훼손, 무단 광고, 불법 정보 유포시 삭제될 수 있으며, 이에 대한 민형사상 책임은 게시자에게 있습니다."
-              maxlength="100"
-            ></textarea>
-            <div class="textarea-footer">
-              <span class="text-counter">0/100</span>
-              <div>
-                <label class="checkbox-label">
-                  <input type="checkbox" class="input-checkbox" />
-                  비밀글 등록
-                </label>
-                <button class="write-btn" type="button">등록하기</button>
-              </div>
-            </div>
-          </div>
-        </form>
-      <button class="order-btn" type="button">
-        최신순
-        <img src="./src/img/down-icon.png" alt="최신순 화살표 아이콘 이미지" />
-      </button>
-      <ul class="comment-lists">
-        <!-- 댓글목록 -->
-      </ul>
-    </main>
-    <article class="modal">
-      <h2 class="a11y-hidden">modal</h2>
-      <div class="dim"></div>
-      <form class="modal-form">
-          <h3 class="modal-title">작성자 확인</h3>
-          <p class="modal-text">비밀번호를 입력해주세요.</p>
-          <input class="input-password" type="password" />
-        <div class="modal-btns">
-          <button type="submit" class="btn-confirm">Confirm</button>
-          <button type="button" class="btn-cancel">Cancel</button>
-        </div>
-    </form>
-    </article>
-    <script src="https://cdn.jsdelivr.net/npm/uuid@8.3.2/dist/umd/uuid.min.js"></script>
-    <script src="./src/js/comment/libray.js" type="module"></script>
-    <script src="./src/js/comment/data.js" type="module"></script>
-    <script src="./src/js/comment/events.js" type="module"></script>
-    <script src="./src/js/comment/render.js" type="module"></script>
-    <script src="./src/js/modal/modal.js" type="module"></script>
-  </body>
-</html>
