--- 230502/230512customPrompt 및 모듈화_남종태/src/style/modal.css
@@ -1,83 +1,0 @@
-.modal {
-  display: none;
-  position: fixed;
-  z-index: 1;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-}
-.dim{
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.4); /* 배경에 어둡게 처리 */
-}
-.modal.active{
-  display: block;
-}
-.modal-form {
-  position: relative;
-  background-color: #fff;
-  margin: 30px auto;
-  padding: 20px;
-  border: 1px solid #888;
-  width: calc(100% - 60px);
-  max-width: 450px;
-  border-radius: 5px;
-}
-.modal-title {
-  font-size: 20px;
-  font-weight: bold;
-  margin-bottom: 20px;
-}
-.modal-text{
-  margin-bottom: 20px;
-  font-size: 14px;
-}
-
-.input-password{
-  width: 100%;
-  padding: 10px;
-  box-sizing: border-box;
-  border-radius: 5px;
-  border: 2px solid #bdbdbd;
-  margin-bottom: 20px;
-}
-.input-password:focus{
-  outline: none;
-  border: 2px solid dodgerblue;
-}
-
-.modal-btns {
-  text-align: right;
-}
-
-.btn-confirm {
-  margin-right: 10px;
-  background-color: #4caf50;
-  color: white;
-  border: none;
-  border-radius: 5px;
-  padding: 10px 15px;
-  cursor: pointer;
-  transition: all 0.5s;
-}
-
-.btn-confirm:hover {
-  background-color: #3e8e41;
-}
-
-.btn-cancel {
-  background-color: #f44336;
-  color: white;
-  border: none;
-  border-radius: 5px;
-  padding: 10px 15px;
-  cursor: pointer;
-  transition: all 0.5s;
-}
-
-.btn-cancel:hover {
-  background-color: #da190b;
-}
