--- 230502/비밀글설정_오연주/README.md
@@ -1,2 +1,0 @@
-# -Comment-writer
-로컬스토리지를 이용한 댓글 작성기
