--- 230502/비밀글설정_오연주/style.css
@@ -1,209 +1,0 @@
-*{
-    padding: 0;
-    margin: 0;
-  }
-  ul{
-    list-style-type: none;
-  }
-  .title {
-    text-align: center;
-    margin: 20px 0px;
-  }
-  .a11y-hidden{
-    clip: rect(1px, 1px, 1px, 1px);
-    clip-path: inset(50%);
-    width: 1px;
-    height: 1px;
-    margin: -1px;
-    overflow: hidden;
-    padding: 0;
-    position: absolute;
-  }
-  
-  main {
-    margin: 0 auto;
-    width: calc(100% - 60px);
-    padding: 30px;
-    max-width: 980px;
-  }
-  section {
-    margin-bottom: 20px;
-  }
-  .section-title{
-    margin-bottom: 20px;
-  }
-  .comment-icon{
-    width: 30px;
-    vertical-align: top;
-  }
-  .comment-form, .edit-form {
-    margin-bottom: 40px;
-  }
-  .comment-form input, .edit-form input{
-    width: 118px;
-    padding: 10px;
-    border: 2px solid #eaeaea;
-    margin-bottom: 20px;
-  }
-  
-  .textarea-wrapper{
-    border: 2px solid #eaeaea;
-  }
-  #input-comment, .input-editComment {
-    width: 100%;
-    height: 100px;
-    resize: none;
-    padding: 20px;
-    box-sizing: border-box;
-    border: none;
-  }
-  .textarea-footer {
-    display: flex;
-    justify-content: space-between;
-    width: 100%;
-    border-top: 2px solid #eaeaea;
-    align-items: center;
-    position: relative;
-  }
-  .text-counter {
-    margin-left: 20px;
-    font-weight: 500;
-    font-size: 16px;
-    color: #bdbdbd;
-  }
-
-    /*비밀글 등록 */
-    #hidden-comment{
-     position: absolute;
-     /* right: 130px; */
-     top:50%;
-      right: 10%;
-      transform:translate(-50%,-50%)
-    
-    }
-    label[for="hidden-comment"] {
-      color: #bdbdbd;
-      font-size: 14px;
-      padding-left: 680px;
-    }
-
-   .text-box{
-    margin-left: 100px;
-   }
-   
-  .comment-box.inactive{
-    display: none;
-  }
-  
-  .edit-form{
-    display: none;
-    border: 2px solid #eee;
-  }
-  .edit-form.active{
-    display: block;
-  }
-  .write-btn, .editComment-btn, .cancel-btn {
-    background-color: gold;
-    height: 52px;
-    width: 91px;
-    cursor: pointer;
-    font-size: 16px;
-    border: none;
-    font-weight: bold;
-  }
-  .editComment-btn{
-    background-color: gold;
-    color: black;
-  }
-  .cancel-btn{
-    background-color: #eee;
-    color: black
-  }
-  
-  .comment-item{
-    border-bottom: 1px solid #bdbdbd;
-    padding-bottom: 15px;
-  }
-  .user-profile{
-    display: flex;
-    align-items: center;
-    margin-bottom: 10px;
-    gap: 8px;
-  }
-  .profile-img{
-    width: 44px;
-    height: 44px;
-    border-radius: 50%;
-  }
-  .auth{
-    margin-left: 3px;
-    font-size: 16px;
-    font-weight: bold;
-  }
-  .comment-item:not(:last-child){
-    margin-bottom: 20px;
-  }
-  .comment-content{
-    margin-bottom: 5px;
-    margin-left: 10px;
-    white-space: pre-wrap;
-  }
-  .createdAt{
-    font-size: 12px;
-    color: #bdbdbd;
-    margin-left: 10px;
-  }
-  .comment-info.inactive{
-    display: none;
-  }
-  
-  .profile-box{
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-  }
-  .edit-comment{
-    display: none;
-    width: 100%;
-    height: 80px;
-    resize: none;
-    padding: 20px;
-    box-sizing: border-box;
-  }
-  .edit-comment.active{
-    display: block;
-  }
-  .edit-btn, .show-btn, .del-btn{
-    width: 30px;
-    height: 30px;
-    border: none;
-    cursor: pointer;
-  }
-  .edit-btn{
-    background: url(./img/edit-icon.png) no-repeat center / 24px;
-  }
-  .show-btn{
-    background: url(./img/show-icon.png) no-repeat center / 24px;
-  }
-  .del-btn {
-    background: url(./img/delete-icon.png) no-repeat center / 24px;
-  }
-  .order-btn{
-    padding: 5px 10px;
-    border: 2px solid #bdbdbd;
-    display: block;
-    margin-left: auto;
-    margin-bottom: 10px;
-    cursor: pointer;
-  }
-  .order-btn img{
-    width: 16px;
-    height: 16px;
-    vertical-align: bottom;
-    transition: all 0.5s;
-  }
-  .order-btn.up img{
-    transform: rotate(-180deg);
-  }
-  
-
